requirements.txt
Flask==2.3.3
python-a2s==1.3.0
gunicorn==21.2.0
python-dotenv==1.0.0 